refactor(api): split fs-based node registry loader from pure builder
Extract `buildNodeRegistry` into a pure module so the registry can be
constructed from data loaded at runtime (e.g. bundled JSON on Workers)
without dragging `node:fs` into environments that don't have it.

- `lib/node-registry.ts` — pure `buildNodeRegistry(data)` + types
- `lib/node-registry-fs.ts` — Node-only `loadNodeRegistry(path)` that
  reads the file and delegates to the pure builder
- Barrel (`src/index.ts`) no longer re-exports the fs variant so non-Node
  consumers can import `createApp` without pulling `node:fs/promises`

Node entries (`serve.ts`, `dev.ts`) import from `node-registry-fs.js`;
behavior unchanged.
