+++ apps/api/src/entry/dev.ts
@@ -9,7 +9,7 @@ import { createApp, createAuth } from '../index.js'
 import { cleanupLocalDevSessions } from '../routes/local-dev.js'
 import { createTokenCrypto } from '../lib/token-crypto.js'
 import { createLogger } from '../lib/logger.js'
-import { loadNodeRegistry } from '../lib/node-registry.js'
+import { loadNodeRegistry } from '../lib/node-registry-fs.js'
 import { createRemoteNodeRegistry } from '../lib/remote-node-registry.js'
 import { createEmailService } from '../lib/email.js'
