+++ apps/api/src/index.ts
@@ -4,5 +4,5 @@ export { createAuth } from './auth/config.js'
 export type { Auth, AuthOptions } from './auth/config.js'
 export { createLogger } from './lib/logger.js'
 export type { Logger } from './lib/logger.js'
-export { loadNodeRegistry } from './lib/node-registry.js'
-export type { AppNodeRegistry } from './lib/node-registry.js'
+export { buildNodeRegistry, createMergedNodeRegistry } from './lib/node-registry.js'
+export type { AppNodeRegistry, RegistryFile } from './lib/node-registry.js'
